refactor: 프로젝트 타입 공통 타입 파일에 분리
